[refactor] Move PureIntrinsic class to intrinsic module. [python/assassyn/ir/expr/intrinsic.py] Add PureIntrinsic class. [python/assassyn/ir/expr/expr.py] Remove PureIntrinsic class. [python/assassyn/ir/expr/__init__.py] Export PureIntrinsic from intrinsic module. [python/assassyn/codegen/simulator/modules.py] Update import. [python/assassyn/codegen/simulator/node_dumper.py] Update import. [python/assassyn/codegen/verilog/design.py] Update import. [python/assassyn/ir/module/base.py] Update import. [python/assassyn/ir/module/module.py] Update import. [python/assassyn/ir/value.py] Update import. [scripts/pre-commit] Add error checking and exit on failure.
