+++ python/assassyn/ir/expr/__init__.py
@@ -2,7 +2,8 @@
 
 #pylint: disable=wildcard-import
 from .expr import *
-from .intrinsic import Intrinsic, finish, wait_until, assume, barrier, mem_write, send_read_request
-from .intrinsic import send_write_request, has_mem_resp, mem_resp, use_dram
+from .intrinsic import Intrinsic, PureIntrinsic, finish, wait_until, assume, barrier
+from .intrinsic import mem_write, send_read_request, send_write_request
+from .intrinsic import has_mem_resp, mem_resp, use_dram
 from .call import Bind, AsyncCall, FIFOPush
 from . import comm
